Fixed random number issues. Fixed compiler warnings.

+++ Ball.cpp
@@ -199,7 +199,7 @@ void Ball::applyPowerUp ( const POWER_UP_TYPE type, ParanoidGameSource * const g
 
 	m_powerUpApplied = true;
 	m_symbol [ 0 ].UnicodeChar = BALL_SYMBOL_PENETRATE;
-	m_powerUpResetTime = game->getGameTime ( ) + POWER_UP_PENETRATION_DURATION;
+	m_powerUpResetTime = static_cast< float >( game->getGameTime ( ) + POWER_UP_PENETRATION_DURATION );
 }
 
 void Ball::resetPowerUps ( )
